requirements.txt
python-dotenv
requests
python-dateutil
zipcodes
twilio
emoji